requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
Pillow==9.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
redis==4.2.2  # https://github.com/andymccurdy/redis-py
#hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==5.0.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-bea
whitenoise==5.3.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==3.1.14  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.42.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.2.0  # https://github.com/jazzband/django-redis
djangorestframework==3.14.0

# Webpack Loader
django-webpack-loader==0.7.0

django-activity-stream==1.4.0

# Extra settings
django-extra-settings==0.7.0

mpmath
numpy